backend/requirements.txt
aioredis==1.3.1
asgiref==3.3.0
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.3
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.0rc1
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-pandas==0.6.4
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
et-xmlfile==1.0.1
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==1.7.0
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
joblib==1.0.1
loguru==0.5.3
MarkupSafe==1.1.1
msgpack==1.0.2
numpy==1.20.2
oauthlib==3.1.0
openpyxl==3.0.7
packaging==20.4
pandas==1.2.4
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.1
scipy==1.6.2
service-identity==18.1.0
six==1.15.0
sklearn==0.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
threadpoolctl==2.1.0
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.25.11
win32-setctime==1.0.3
zipp==3.4.0
zope.interface==5.4.0
